More php

+++ inc/main.php
@@ -26,7 +26,7 @@
 				 echo "Upload failed.";
 			 }
 		 }else{
-			//Deal with non-csv files
+			//Deal with non-csv files or no files at all
 			echo "Invalid file, please try again... <script type='text/javascipt'>console.log('Invalid File Upload in php');</script>";
 		 }
 	
@@ -43,17 +43,22 @@
 		//Create array holding each line of text from csv file
 		$fileLines = explode("\n\r", $fileContent);
 		//Create multidimensional array for each cell in the file
-		$i = 0;
-		while($i < array_count_values($fileLines)){
-			//As long as there are still more lines in the file to iterate through...
-			$fileCells = explode(",", $fileLines[i]);
-			//Put results into console for debugging
-			echo "<script type='text/javascript'>console.log(".$fileCells.")</script>";
+		if(array_key_exists($_POST[]) || array_key_exists($_FILES['file'])){
+			$i = 0;
+			while($i < array_count_values($fileLines)){
+				//As long as there are still more lines in the file to iterate through...
+				$fileCells = explode(",", $fileLines[i]);
+				//Put results into console for debugging
+				echo "<script type='text/javascript'>console.log(".$fileCells.")</script>";
+			}
+			//Data is already in the necessary format and now I just need to echo it in order for the AJAX call to get it
+			if($_GET['action'] == "printGraph"){
+				exit();
+				echo $data;
+			}
+		}else{
+		 echo "No file detected...";
 		}
-		//Data is already in the necessary format and now I just need to echo it in order for the AJAX call to get it
-		if($_GET['action'] == "printGraph"){
-			exit();
-			echo $data;
-		}
+		
 	
 ?>
